backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "yahoo"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
pvrecorder = "^1.2.2"
datasets = "^2.19.0"
transformers = "^4.40.1"
torch = "^2.3.0"
tweepy = "^4.14.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

